rename nns_dapp module to bootstrap
